Code cleanup and documentation #256 #623
