+++ src/main/groovy/nextflow/processor/ProcessConfig.groovy
@@ -22,8 +22,12 @@ package nextflow.processor
 
 import static nextflow.util.CacheHelper.*
 
+import java.util.regex.Pattern
+
 import groovy.transform.PackageScope
 import groovy.util.logging.Slf4j
+import nextflow.Const
+import nextflow.exception.ConfigParseException
 import nextflow.exception.IllegalConfigException
 import nextflow.exception.IllegalDirectiveException
 import nextflow.executor.BashWrapperBuilder
@@ -51,10 +55,6 @@ import nextflow.script.ValueOutParam
  */
 @Slf4j
 class ProcessConfig implements Map<String,Object> {
-
-    static final public transient BOOL_YES = ['true','yes','on']
-
-    static final public transient BOOL_NO = ['false','no','off']
 
     static final public transient LABEL_REGEXP = ~/[a-zA-Z]([a-zA-Z0-9_]*[a-zA-Z0-9]+)?/
 
@@ -101,8 +101,16 @@ class ProcessConfig implements Map<String,Object> {
             'stageOutMode'
     ]
 
+    /**
+     * Names of directives that can be used more than once in the process definition
+     */
+    @PackageScope
     static final List<String> repeatableDirectives = ['label','module','publishDir']
 
+    /**
+     * Default directives values
+     */
+    @PackageScope
     static final Map<String,Object> DEFAULT_CONFIG = [
             echo: false,
             cacheable: true,
@@ -113,23 +121,43 @@ class ProcessConfig implements Map<String,Object> {
             errorStrategy: ErrorStrategy.TERMINATE
     ]
 
+    /**
+     * Map instance that holds the process configuration
+     */
     @Delegate
+    @PackageScope
     protected final Map<String,Object> configProperties
 
+    /**
+     * Reference to the main script object
+     */
     private BaseScript ownerScript
 
+    /**
+     * Name of the process to which this config object is associated
+     */
+    private String processName
+
+    /**
+     * When {@code true} a {@link MissingPropertyException} is thrown when
+     * trying to access a property not existing
+     */
     private boolean throwExceptionOnMissingProperty
 
+    /**
+     * List of process input definitions
+     */
     private inputs = new InputsList()
 
+    /**
+     * List of process output definitions
+     */
     private outputs = new OutputsList()
 
     /**
      * Initialize the taskConfig object with the defaults values
      *
      * @param script The owner {@code BaseScript} configuration object
-     * @param important The values specified by this map won't be overridden by attributes
-     *      having the same name defined at the task level
      */
     ProcessConfig( BaseScript script ) {
         ownerScript = script
@@ -138,18 +166,23 @@ class ProcessConfig implements Map<String,Object> {
     }
 
     /* Only for testing purpose */
-    protected ProcessConfig( Map delegate ) {
+    @PackageScope
+    ProcessConfig( Map delegate ) {
         configProperties = delegate
     }
 
+    /**
+     * Define the name of the process to which this config object is associated
+     *
+     * @param name String representing the name of the process to which this config object is associated
+     * @return The {@link ProcessConfig} instance itself
+     */
     @PackageScope
-    static boolean toBool( value )  {
-        if( value instanceof Boolean ) {
-            return value.booleanValue()
-        }
-
-        return value != null && value.toString().toLowerCase() in BOOL_YES
+    ProcessConfig setProcessName( String name ) {
+        this.processName = name
+        return this
     }
+
 
     /**
      * Enable special behavior to allow the configuration object
@@ -159,7 +192,7 @@ class ProcessConfig implements Map<String,Object> {
      * @return The object itself
      */
     @PackageScope
-    ProcessConfig enterCaptureMode( boolean value ) {
+    ProcessConfig throwExceptionOnMissingProperty( boolean value ) {
         this.throwExceptionOnMissingProperty = value
         return this
     }
@@ -212,7 +245,7 @@ class ProcessConfig implements Map<String,Object> {
         return this
     }
 
-    def getProperty( String name ) {
+    Object getProperty( String name ) {
 
         switch( name ) {
             case 'inputs':
@@ -263,6 +296,120 @@ class ProcessConfig implements Map<String,Object> {
     }
 
     /**
+     * Apply the settings defined in the configuration file for the given annotation label, for example:
+     *
+     * ```
+     * process {
+     *     withLabel: foo {
+     *         cpus = 1
+     *         memory = 2.gb
+     *     }
+     * }
+     * ```
+     *
+     * @param configDirectives
+     *      A map object modelling the setting defined defined by the user in the nextflow configuration file     *
+     * @param category
+     *      The type of annotation either {@code withLabel:} or {@code withName:}
+     * @param processLabel
+     *      A specific label representing the object holding the configuration setting to apply
+     */
+    @PackageScope
+    void applyConfigForLabel( Map<String,?> configDirectives, String category, String processLabel ) {
+        assert category in ['withLabel:','withName:']
+        assert processName != null
+        assert processLabel != null
+
+        for( String rule : configDirectives.keySet() ) {
+            if( !rule.startsWith(category) )
+                continue
+            def isLabel = category=='withLabel:'
+            def pattern = rule.substring(category.size()).trim()
+            final isNegated = pattern.startsWith('!')
+            if( isNegated )
+                pattern = pattern.substring(1).trim()
+            final target = isLabel ? processLabel : processName
+            final matches = Pattern.compile(pattern).matcher(target).matches() ^ isNegated
+            if( !matches )
+                continue
+
+            log.debug "Config settings `$rule` matches ${isLabel ? "label `$target` for process with name $processName" : "process $processName"}"
+            def settings = configDirectives.get(rule)
+            if( settings instanceof Map ) {
+                applyConfigSettings(settings)
+            }
+            else if( settings != null ) {
+                throw new ConfigParseException("Unknown config settings for label `$processLabel` -- settings=$settings ")
+            }
+        }
+    }
+
+    /**
+     * Apply the settings defined in the configuration file to the actual process configuration object
+     *
+     * @param settings
+     *      A map object modelling the setting defined defined by the user in the nextflow configuration file
+     */
+    @PackageScope
+    void applyConfigSettings(Map<String,?> settings) {
+        if( !settings )
+            return
+
+        for( Entry<String,?> entry: settings ) {
+            if( entry.key.startsWith('$'))
+                continue
+
+            if( entry.key.startsWith("withLabel:") || entry.key.startsWith("withName:"))
+                continue
+
+            if( !DIRECTIVES.contains(entry.key) )
+                log.warn "Unknown directive `$entry.key` for process `$processName`"
+
+            if( entry.key == 'params' ) // <-- patch issue #242
+                continue
+
+            if( entry.key == 'ext' ) {
+                if( this.getProperty('ext') instanceof Map ) {
+                    // update missing 'ext' properties found in 'process' scope
+                    def ext = this.getProperty('ext') as Map
+                    entry.value.each { String k, v -> ext[k] = v }
+                }
+                continue
+            }
+
+            this.put(entry.key,entry.value)
+        }
+    }
+
+    /**
+     * Apply the process settings defined globally in the process config scope
+     *
+     * @param processDefaults
+     *      A map object representing the setting to be applied to the process
+     *      (provided it does not already define a different value for
+     *      the same config setting).
+     *
+     */
+    @PackageScope
+    void applyConfigDefaults( Map processDefaults ) {
+        for( String key : processDefaults.keySet() ) {
+            if( key == 'params' )
+                continue
+            def value = processDefaults.get(key)
+            def current = this.get(key)
+            if( key == 'ext' ) {
+                if( value instanceof Map && current instanceof Map ) {
+                    def ext = current as Map
+                    value.each { k,v -> if(!ext.containsKey(k)) ext.put(k,v) }
+                }
+            }
+            else if( current==null || current == ProcessConfig.DEFAULT_CONFIG.get(key) ) {
+                this.put( key, value )
+            }
+        }
+    }
+
+    /**
      * Type shortcut to {@code #configProperties.inputs}
      */
     InputsList getInputs() {
@@ -275,7 +422,6 @@ class ProcessConfig implements Map<String,Object> {
     OutputsList getOutputs() {
         outputs
     }
-
 
     /*
      * note: without this method definition {@link BaseScript#echo} will be invoked
@@ -340,7 +486,6 @@ class ProcessConfig implements Map<String,Object> {
         result
     }
 
-
     /**
      * Defines a special *dummy* input parameter, when no inputs are
      * provided by the user for the current task
@@ -353,7 +498,6 @@ class ProcessConfig implements Map<String,Object> {
         new DefaultOutParam(this)
     }
 
-
     boolean isCacheable() {
         def value = configProperties.cache
         if( value == null )
@@ -362,7 +506,7 @@ class ProcessConfig implements Map<String,Object> {
         if( value instanceof Boolean )
             return value
 
-        if( value instanceof String && value in BOOL_NO )
+        if( value instanceof String && value in Const.BOOL_NO )
             return false
 
         return true
@@ -382,6 +526,17 @@ class ProcessConfig implements Map<String,Object> {
         return LABEL_REGEXP.matcher(left).matches() && LABEL_REGEXP.matcher(right).matches()
     }
 
+    /**
+     * Implements the process {@code label} directive.
+     *
+     * Note this directive  can be specified (invoked) more than one time in
+     * the process context.
+     *
+     * @param lbl
+     *      The label to be attached to the process.
+     * @return
+     *      The {@link ProcessConfig} instance itself.
+     */
     ProcessConfig label(String lbl) {
         if( !lbl ) return this
         // -- check that label has a valid syntax
@@ -405,6 +560,16 @@ class ProcessConfig implements Map<String,Object> {
         (List<String>) configProperties.get('label') ?: Collections.emptyList()
     }
 
+    /**
+     * Implements the process {@code module} directive.
+     *
+     * See also http://modules.sourceforge.net
+     *
+     * @param moduleName
+     *      The module name to be used to execute the process.
+     * @return
+     *      The {@link ProcessConfig} instance itself.
+     */
     ProcessConfig module( moduleName ) {
         // when no name is provided, just exit
         if( !moduleName )
@@ -423,9 +588,18 @@ class ProcessConfig implements Map<String,Object> {
         return this
     }
 
-
-    ProcessConfig errorStrategy( value ) {
-        configProperties.put('errorStrategy', value)
+    /**
+     * Implements the {@code errorStrategy} directive
+     *
+     * @see ErrorStrategy
+     *
+     * @param strategy
+     *      A string representing the error strategy to be used.
+     * @return
+     *      The {@link ProcessConfig} instance itself.
+     */
+    ProcessConfig errorStrategy( strategy ) {
+        configProperties.put('errorStrategy', strategy)
         return this
     }
