+++ src/test/groovy/nextflow/executor/BashWrapperBuilderTest.groovy
@@ -1938,7 +1938,6 @@ class BashWrapperBuilderTest extends Specification {
 
                 nxf_mktemp() {
                     local base=\${1:-/tmp}
-                    if [[ \$base == /dev/shm && ! -d \$base ]]; then base=/tmp; fi 
                     if [[ \$(uname) = Darwin ]]; then mktemp -d \$base/nxf.XXXXXXXXXX
                     else TMPDIR="\$base" mktemp -d -t nxf.XXXXXXXXXX
                     fi
@@ -1971,7 +1970,7 @@ class BashWrapperBuilderTest extends Specification {
                 [[ \$NXF_SCRATCH ]] && echo "nxf-scratch-dir \$HOSTNAME:\$NXF_SCRATCH" && cd \$NXF_SCRATCH
 
                 set +e
-                ctmp=\$(nxf_mktemp /dev/shm)
+                ctmp=\$(set +u; nxf_mktemp /dev/shm 2>/dev/null || nxf_mktemp \$TMPDIR)
                 cout=\$ctmp/.command.out; mkfifo \$cout
                 cerr=\$ctmp/.command.err; mkfifo \$cerr
                 tee .command.out < \$cout &
